spa/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-common-helpers==0.9.2
django-cron==0.5.1
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
